adds sitemap index generation and configuration options

+++ backend/job_runners/aspace_sitemap_runner.rb
@@ -1,11 +1,19 @@
+require 'zip'
+require 'date'
+
 class AspaceSitemapRunner < JobRunner
   
   register_for_job_type('aspace_sitemap_job')
   
   def run
+    # these could become user set in the frontend, but not right now
+    sitemap_limit = AppConfig.has_key?(:aspace_sitemap_limit) ? AppConfig[:aspace_sitemap_limit] : 50000
+    sitemap_index_base_url = AppConfig.has_key?(:aspace_sitemap_baseurl) ? AppConfig[:aspace_sitemap_baseurl] : "https://change.to.my.site.baseurl/"
+    timestamp = Time.now.strftime("%Y-%d-%m-%H-%M-%S-%L")
+
     @job.write_output('Generating sitemap')
-    file = ASUtils.tempfile('aspace_sitemap_job_')
     array = []
+    files = []
     begin
       DB.open do |db|
         db.fetch(query_string) do |result|
@@ -15,30 +23,76 @@ class AspaceSitemapRunner < JobRunner
         end
       end
       
-      builder = Nokogiri::XML::Builder.new(:encoding => 'UTF-8') do |xml|
-        xml.urlset('xmlns' => "https://www.sitemaps.org/schemas/sitemap/0.9") {
-          array.each do |entry|
-            xml.url {
-              xml.loc entry[:loc]
-              xml.lastmod entry[:lastmod]
-              xml.changefreq entry[:changefreq]
-              xml.priority entry[:priority]
+      # split the results set into chunks of less than the sitemap entry limit
+      sitemap_parts = array.each_slice(sitemap_limit).to_a
+
+      # initialize a zip file
+      zip_file = ASUtils.tempfile("aspace_sitemap_zip_#{timestamp}")
+      # open with the OutputStream class to initialize the zip struture correctly
+      Zip::OutputStream.open(zip_file) { |zos| }
+
+      # iterate through the sitemap piecesand build our xml files
+      sitemap_parts.each_with_index do |sitemap,k|
+        files[k] = Tempfile.new(["aspace_sitemap_#{timestamp}_part_#{k}", ".xml"])
+      
+        builder = Nokogiri::XML::Builder.new(:encoding => 'UTF-8') do |xml|
+          xml.urlset('xmlns' => "https://www.sitemaps.org/schemas/sitemap/0.9") {
+            sitemap.each do |entry|
+              xml.url {
+                xml.loc entry[:loc]
+                xml.lastmod entry[:lastmod]
+                xml.changefreq entry[:changefreq]
+                xml.priority entry[:priority]
+              }
+            end
+          }
+        end
+        files[k].write(builder.to_xml)
+        files[k].rewind
+      end
+      
+      # create a sitemap index file if necessary
+      if sitemap_parts.count > 1
+        index_file = Tempfile.new(["aspace_sitemap_index_#{timestamp}",".xml"])
+
+        index_builder = Nokogiri::XML::Builder.new(:encoding => 'UTF-8') do |xml|
+            xml.urlset('xmlns' => "https://www.sitemaps.org/schemas/sitemap/0.9") {
+              files.each_with_index do |file,k|
+                xml.sitemap {
+                  xml.loc "#{sitemap_index_base_url}aspace_sitemap_#{timestamp}_part_#{k}.xml"
+                  xml.lastmod timestamp
+                }
+              end
             }
           end
-        }
+        index_file.write(index_builder.to_xml)
+        index_file.rewind
       end
-      file.write(builder.to_xml)
-      file.rewind
+      
+      # wrap them all into a zip file
+      Zip::File.open(zip_file.path, Zip::File::CREATE) do |zip|
+        files.each_with_index do |file,k|
+          zip.add("aspace_sitemap_#{timestamp}_part_#{k}.xml", file.path)
+        end
+        zip.add("aspace_sitemap_index_#{timestamp}.xml", index_file.path) if sitemap_parts.count > 1
+      end
+      
       @job.write_output('Adding Sitemap')
-      @job.add_file(file)
+      @job.add_file(zip_file)
       self.success!
     rescue Exception => e
       @job.write_output(e.message)
       @job.write_output(e.backtrace)
       raise e
     ensure
-      file.close
-      file.unlink
+      files.each do |file|
+        file.close
+        file.unlink
+      end
+      index_file.close if sitemap_parts.count > 1
+      index_file.unlink if sitemap_parts.count > 1
+      zip_file.close
+      zip_file.unlink
       @job.write_output('Done.')
     end
   end
@@ -50,7 +104,7 @@ class AspaceSitemapRunner < JobRunner
       row[:loc] = ["#{AppConfig[:public_proxy_url]}","repositories",row[:repo_id],row[:source],row[:id]].join("/")
     end
     row[:lastmod] = row[:lastmod].strftime("%Y-%m-%d")
-    row[:changefreq] = "yearly"
+    row[:changefreq] = AppConfig.has_key?(:aspace_sitemap_changefreq)? AppConfig[:aspace_sitemap_changefreq] : "yearly"
     row[:priority] = "0.5"
     row.delete(:id)
     row.delete(:repo_id)
